Fix printing QList<bool> contents
Internally, a QList<T> stores items as a QList<T>::Node. These are
assigned via a pointer (see QList<T>::node_construct) and must also be
read via a pointer dereference. Failing to do so results in unexpected
values because non-zero is a truth value:
because of the following:

    v = 0x7fffffffe300
    sizeof(bool) == 1
    (bool)v == true
    *(bool *)v == false

Test:

    cat >test.cpp <<EOF
    #include <QList>
    int main(void) {
        QList<bool> list;
        list.append(false);
        list.append(false);
        asm("int3");
        return 0;
    }
    EOF
    g++ -fPIC test.cpp `pkg-config --cflags --libs Qt5Core` -g &&
    gdb -q ./a.out -batch -ex r -ex p\ list

Expected result:

    $1 = {false, false}

Actual result:

    $1 = {false, true}